Add numpy-based PageRank fallback and fix Neo4j query count
